fix preset button, refactor, fix ccm
