+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/content/guide/PylonGuide.kt
@@ -24,6 +24,7 @@ class PylonGuide(stack: ItemStack) : PylonItem(stack), PylonInteractor {
     override fun onUsedToRightClick(event: PlayerInteractEvent) {
         if (event.action.isRightClick) {
             open(event.player)
+            event.isCancelled = true
         }
     }
 
@@ -33,7 +34,7 @@ class PylonGuide(stack: ItemStack) : PylonItem(stack), PylonInteractor {
         val KEY = pylonKey("guide")
 
         @JvmField
-        val STACK = ItemStackBuilder.Companion.pylonItem(Material.ENCHANTED_BOOK, KEY)
+        val STACK = ItemStackBuilder.pylonItem(Material.KNOWLEDGE_BOOK, KEY)
             .build()
 
         /**
